document v1.1: support inactive members
The original driver of this was that I wanted to have the ability
to put last year's list into the config as rules (so that nobody
draws the same name as last year) but also retain the ability to
comment out the names of folks that aren't participating this year.

I played around with making the validation errors non-fatal, but
introducing the concept of "inactivity" feels more robust.